Parameterize generation and logging
Instead of putting hard constraints on Arbitrary, ToJSON, FromJSON and
other type classes we can full parameterize with ConstraintKinds. This
brings a lot more weight to the `Graph` type signature but it makes all
of the constraints fully pluggable. With this you basically get a single
algebra for constructing graphs with pluggable frontends and backends.