Switch graphula functions to utilize MonadGraphula
We likely don't need the free monad implementation of graphula. It was
initially made for prototyping purposes, because we were unsure if we
could accomplish all necessary goals without it. It seems goals can be
accomplished with a final tagless representation. This commit introduces
MonadGraphulaFrontend and MonadGraphulaBackend, the two demarcations of
functionality in Graphula. We can likely split this even more, but it
isn't necessary now.