Add constraint based editing
`suchThat` is a powerful tool in property based testing. Allowing for
patterns like `suchThat` enables more powerful testing with `graphula`.
This is a proof of concept that proposes changing `edit` to a combinator
of type `a -> Maybe a` instead of `a -> a`. With the addition of `Maybe`
we can allow for failure in an edit, which leads to constraint based
combinators like `suchThat`.

```hs
edit :: (a -> a) -> a -> Maybe a
edit f = Just . f

suchThat :: (a -> Bool) -> a -> Maybe a
suchThat f x = x <$ guard (f x)
```