Add emplaceKey to EntityKeyGen
When a node implements EntityKeyGen such that genEntityKey returns
Just, we use insertKey to write the node to the database. However,
if the key is a composite key, persistent keeps those values in the
record itself, and insertKey doesn't ensure that the key argument and
the record agree on those values.

This is bad.

I've added emplaceKey as a method on EntityKeyGen with a default
implementation, which just returns the record unchanged. There is also a
generic implementation for composite keys called emplaceCompositeKey.
Unfortunately, I can't use this as the default because regular Keys
don't have a Generic instance because they wrap a BackendKey.