Simplify node API while expanding the set of supported relationships
Previously, we had:

  node         :: (Dependencies a ~ ())             => m (Entity a)
  nodeEdit     :: (Dependencies a ~ ()) => (a -> a) -> m (Entity a)
  nodeWith     ::  Dependencies a                   -> m (Entity a)
  nodeEditWith ::  Dependencies a       -> (a -> a) -> m (Entity a)

We wanted to add a way to override an entity's primary key at the call
site, but that would necessarily expand the set of bindings we'd have to
export. So, I've changed the API to just:

  root         :: (Dependencies a ~ ()) => NodeOptions a -> m (Entity a)
  node         ::  Dependencies a       -> NodeOptions a -> m (Entity a)

`NodeOptions a` is an opaque data type constructed with:

  mempty       ::                NodeOptions a
  edit         :: (a -> a)    -> NodeOptions a
  suchThat     :: (a -> Bool) -> NodeOptions a
  primaryKey   :: Key a       -> NodeOptions a

Multiple `NodeOptions a` values can be combined with the `Semigroup`
operator `(<>)`:

  Entity schoolId _ <- root @School
    $ primaryKey (SchoolKey 1)
    <> edit (\s -> s { schoolName = "The School" })

One can avoid the extra parenthesization around the lambda with `do`:

  Entity schoolId _ <- root @School
    $ primaryKey (SchoolKey 1)
    <> do edit $ \s -> s { schoolName = "The School" }

Or by using `mconcat`:

  Entity schoolId _ <- root @School $ mconcat
    [ primaryKey (SchoolKey 1)
    , edit $ \s -> s { schoolName = "The School" }
    ]

I don't expect the `primaryKey` option to be used very often, so in the
usual case, we'd mostly likely be using `mempty` or `edit` alone.