Specify KeySource in HasDependencies
This simplifies the `node` API even further. For a regular entity, a
client now declares:

  mkPersist sqlSettings [persistLowerCase|
    Teacher
      name Text
      deriving Eq Show Generic

    Student
      name Text
      teacherId TeacherId Maybe
      deriving Eq Show Generic
  |]

  instance HasDependencies Teacher where
    type Dependencies Teacher = ()                 -- default
    type KeySource Teacher = 'GenerateKey 'Default -- default

  instance HasDependencies Student where
    type Dependencies Student = Only (Maybe TeacherId)
    type KeySource Student = 'GenerateKey 'Default -- default

Using this information, we can statically assert that these values
will always get their key from the database. We can use `node` as before
to create entities:

  Entity teacherId _ <- node @Teacher $ edit $ \t -> t { teacherName = "Joe" }
  Entity studentId _ <- node @Student (only $ Just teacherId) mempty

If we have a type with key that can be generate with `Arbitrary`, we can
do like so:

  mkPersist sqlSettings [persistLowerCase|
    MathQuestion
      Id QuestionIdText
      skillId SkillId
      answerType AnswerType
      deriving Eq Show Generic
  |]

  deriving instance {-# OVERLAPPING #-} Arbitrary (Key MathQuestion)

  instance HasDependencies MathQuestion where
    type Dependencies MathQuestion = Only SkillId
    type KeySource MathQuestion = 'GenerateKey 'Arbitrary

Again, `node` works as before, only we'll use `Arbitrary` ahead of time
to create a key:

  Entity questionId _ <- node @MathQuestion (only skillId) mempty

Finally, if we have a type whose key should always be externally
supplied, we do like so:

  mkPersist sqlSettings [persistLowerCase|
    MathQuestionContent
      Id QuestionIdText
      content Content
      deriving Eq Show Generic
  |]

  instance HasDependencies MathQuestionContent where
    type Dependencies MathQuestionContent = ()        -- default
    type KeySource MathQuestionContent = 'SpecifyKey

`node` will require an extra argument after the type's dependencies:

  Entity questionId _ <- node @MathQuestion (only skillId) mempty
  Entity contentId _ <- node @MathQuestionContent () (MathQuestionContentKey questionId) mempty

This uses some magic, but it's hopefully mostly hidden.

Also `GenEntityKey` is dead. `'GenerateKey 'Arbitrary` makes it redundant.