Remove Replay functionality
Requiring that node types have JSON instances is problematic for us for
nuanced, domain-specific reasons. By removing replay functionality,
which we don't use, we believe we can then remove the JSON requirement
by logging graphs, which we do use, as a serialization format that need
not support de-serialization (e.g. Show).

Ideally, this would be more directly configurable by end-users, such
that anyone who wanted Replay could accept the JSON requirement, without
forcing it on those that don't. But that is deferred for now since we
are the only end users, as far as we know.