Build on nightly
Newer GHC adds two warnings which we can only disable on nightly, since
they're unknown options to older GHCs. Therefore, they were added
directly on the CLI in CI.

Newer GHC (or Cabal?) also adds the unnecessary package warning, which I
think is very valuable so we can't disable it to support the
markdown-unlit executable dependency required for the readme spec. To
avoid that, we have to build it explicitly on CI. Local development will
also hit an error unless it's installed explicitly there. We should
investigate how to smooth this -- e.g. bring back make setup?

Remove askUnliftIO, as it's no longer part of the type class in newer
unliftio-core. NOTE: this is not a compatibility issue since that class
member was always optional anyway.

Unused type variables are now also flagged. Replacing them with
underscores solves this, but *just* an "_" is a syntax error in older
GHCs, so we have to use "_a", even though the "a" part adds no
information.