Build on nightly
Removes askUnliftIO, as it's no longer part of the type class in newer
unliftio-core. NOTE: this is not a compatibility issue since that class
member was always optional anyway.

Unused type variables are now also flagged. Replacing them with
underscores solves this, but *just* an "_" is a syntax error in older
GHCs, so we have to use "_a", even though the "a" part adds no
information.